+++ tests/mod.rs
@@ -1,5 +1,6 @@
-pub mod common;
-pub mod iif;
+mod builtin;
+mod common;
+mod iif;
 
 use crate::common::eval_expr;
 use common::{simple_event, Event};